+++ lib/legion/llm/pipeline.rb
@@ -7,8 +7,8 @@ require_relative 'pipeline/tracing'
 require_relative 'pipeline/timeline'
 require_relative 'pipeline/audit_publisher'
 require_relative 'pipeline/gaia_caller'
+require_relative 'pipeline/tool_dispatcher'
 require_relative 'pipeline/steps'
-require_relative 'pipeline/steps/metering'
 require_relative 'pipeline/executor'
 
 module Legion